backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings>=2.0
python-dotenv
PyJWT>=2.0
httpx
openai>=1.0
pypdf>=4.0
python-docx>=1.0
python-multipart
stripe>=8.0